Removed auto set tab; Added row count and column reordering
Removed the feature that auto sets the active tab dependent on the data
in the tab. Jason pointed out that it was overkill and might confuse the user.

To replace that feature, the tabs now list the number of rows next to the tab name.
This allows the user to quickly identify whether their search returned any data in any
of the tabs.

Also added the ability for the user to reorganize the table columns, using an extension.